+++ omega_logger/main.py
@@ -1,40 +1,23 @@
 """
-Start all OMEGA loggers and the web server.
+Start all OMEGA loggers and the web server or perform a database backup.
 
 Usage:
 
 omega-logger /path/to/config.xml
+omega-logger /path/to/config.xml --backup
 """
 import os
 import re
 import sys
 import time
+import logging
+import sqlite3
 
+from msl.io import search
 from msl.equipment import Config
 
 
-def start():
-    if len(sys.argv) == 1:
-        print('You must pass in the path to the XML configuration file.', file=sys.stderr)
-        return 1
-
-    xml = os.path.abspath(sys.argv[1])
-    try:
-        cfg = Config(xml)
-    except OSError as e:
-        print(f'{e.__class__.__name__}: {e}', file=sys.stderr)
-        return 1
-
-    log_dir = cfg.value('log_dir')
-    if not log_dir:
-        print('There is no "log_dir" element in the config file.\n'
-              'What directory do you want to log the data to?', file=sys.stderr)
-        return 1
-
-    if not os.path.isdir(log_dir):
-        print(f'The log_dir value of {log_dir!r} is not a valid directory.', file=sys.stderr)
-        return 1
-
+def run_webapp(cfg, xml):
     serials = cfg.value('serials')
     if not serials:
         print('You have not specified a serial number of an OMEGA iServer.\n'
@@ -103,3 +86,102 @@ def start():
     # start the Dash web application
     cmd = prefix + ' '.join([sys.executable, '-m', 'webapp', f'"{xml}"', ','.join(serials)])
     os.system(cmd)
+
+
+def run_backup(log_dir, backup_dir=None):
+    if sys.version_info[:2] < (3, 7):
+        # the sqlite3.Connection.backup() method was added in Python 3.7
+        print('Can only perform a database backup with Python 3.7 or later', file=sys.stderr)
+        return 1
+
+    if not backup_dir:
+        backup_dir = os.path.join(log_dir, 'backup')
+
+    os.makedirs(backup_dir, exist_ok=True)
+
+    # set up logging
+    formatter = logging.Formatter('%(asctime)s [%(levelname)-5s] %(message)s')
+    formatter.default_msec_format = '%s.%03d'
+    file_handler = logging.FileHandler(os.path.join(backup_dir, 'log.txt'))
+    file_handler.setFormatter(formatter)
+    stream_handler = logging.StreamHandler(sys.stdout)
+    stream_handler.setFormatter(formatter)
+    logger = logging.getLogger(__name__)
+    logger.addHandler(file_handler)
+    logger.addHandler(stream_handler)
+    logger.setLevel(logging.INFO)
+
+    logger.info('----- START  BACKUP -----')
+    for file in search(log_dir, pattern=r'\.sqlite3$'):
+        original = sqlite3.connect(file)
+        logger.info(f'processing {os.path.basename(file)}')
+
+        # make sure that the database is not corrupt
+        check = original.execute('PRAGMA integrity_check;').fetchall()
+        if check != [('ok',)]:
+            corrupt = '\n  '.join(item for row in check for item in row)
+            logger.error(f'integrity check failed, database is corrupt:\n  {corrupt}')
+            original.close()
+            continue
+        logger.info('integrity check passed')
+
+        # create the backup
+        backup = sqlite3.connect(os.path.join(backup_dir, os.path.basename(file)))
+        try:
+            original.backup(backup)
+        except sqlite3.Error as e:
+            logger.exception(e)
+            continue
+        logger.info('created backup')
+
+        # verify backup
+        cursor = backup.execute('SELECT * FROM data;')
+        for record in original.execute('SELECT * FROM data;'):
+            fetched = cursor.fetchone()
+            if record != fetched:
+                logger.error(
+                    f'verifying backup failed, record mismatch:'
+                    f'\n  original={record}'
+                    f'\n    backup={fetched}'
+                )
+                backup.close()
+                original.close()
+                break
+
+        if cursor.fetchone() is None:
+            logger.info('verified backup')
+        else:
+            logger.error('verifying backup failed, database size mismatch')
+
+        backup.close()
+        original.close()
+
+    logger.info('----- FINISH BACKUP -----')
+
+
+def start():
+    if len(sys.argv) == 1:
+        print('You must pass in the path to the XML configuration file.', file=sys.stderr)
+        return 1
+
+    xml = os.path.abspath(sys.argv[1])
+    try:
+        cfg = Config(xml)
+    except OSError as e:
+        print(f'{e.__class__.__name__}: {e}', file=sys.stderr)
+        return 1
+
+    log_dir = cfg.value('log_dir')
+    if not log_dir:
+        print('There is no "log_dir" element in the config file.\n'
+              'What directory do you want to log the data to?', file=sys.stderr)
+        return 1
+
+    if not os.path.isdir(log_dir):
+        print(f'The log_dir value of {log_dir!r} is not a valid directory.', file=sys.stderr)
+        return 1
+
+    if '--backup' in sys.argv:
+        return run_backup(log_dir, cfg.value('backup_dir'))
+
+    return run_webapp(cfg, xml)