+++ tests/test_cli.py
@@ -1,4 +1,9 @@
+import os
+import sys
+import shutil
 from subprocess import run, PIPE
+
+import pytest
 
 
 def test_no_args():
@@ -48,3 +53,70 @@ def test_no_connections():
     assert process.returncode == 1
     assert not process.stdout
     assert process.stderr.startswith(b'You have not specified a "connections/connection/path" element')
+
+
+@pytest.mark.skipif(sys.version_info[:2] < (3, 7), reason='requires Python 3.7+')
+def test_backup_default_dir():
+    # the <backup_dir> XML element is not specified in config_backup.xml
+    # so the default backup_dir is used
+    backup_dir = os.path.join('tests', 'resources', 'backup')
+    if os.path.isdir(backup_dir):
+        shutil.rmtree(backup_dir)
+
+    process = run(['omega-logger', 'tests/resources/config_backup.xml', '--backup'], stderr=PIPE, stdout=PIPE)
+    assert process.returncode == 0
+    assert not process.stderr
+    lines = [line[24:] for line in process.stdout.decode().splitlines()]  # ignore the datetime
+    assert lines == [
+        '[INFO ] ----- START  BACKUP -----',
+        '[INFO ] processing iTHX-W3-5_01234.sqlite3',
+        '[INFO ] integrity check passed',
+        '[INFO ] created backup',
+        '[INFO ] verified backup',
+        '[INFO ] processing iTHX-W_56789.sqlite3',
+        '[INFO ] integrity check passed',
+        '[INFO ] created backup',
+        '[INFO ] verified backup',
+        '[INFO ] ----- FINISH BACKUP -----'
+    ]
+    assert os.path.isfile(os.path.join(backup_dir, 'iTHX-W3-5_01234.sqlite3'))
+    assert os.path.isfile(os.path.join(backup_dir, 'iTHX-W_56789.sqlite3'))
+    assert os.path.isfile(os.path.join(backup_dir, 'log.txt'))
+    shutil.rmtree(backup_dir)
+
+
+@pytest.mark.skipif(sys.version_info[:2] < (3, 7), reason='requires Python 3.7+')
+def test_backup_dir():
+    # the <backup_dir> XML element is specified in config_backup2.xml
+    backup_dir = os.path.join('tests', 'resources', 'temp')
+    if os.path.isdir(backup_dir):
+        shutil.rmtree(backup_dir)
+
+    process = run(['omega-logger', 'tests/resources/config_backup2.xml', '--backup'], stderr=PIPE, stdout=PIPE)
+    assert process.returncode == 0
+    assert not process.stderr
+    lines = [line[24:] for line in process.stdout.decode().splitlines()]  # ignore the datetime
+    assert lines == [
+        '[INFO ] ----- START  BACKUP -----',
+        '[INFO ] processing iTHX-W3-5_01234.sqlite3',
+        '[INFO ] integrity check passed',
+        '[INFO ] created backup',
+        '[INFO ] verified backup',
+        '[INFO ] processing iTHX-W_56789.sqlite3',
+        '[INFO ] integrity check passed',
+        '[INFO ] created backup',
+        '[INFO ] verified backup',
+        '[INFO ] ----- FINISH BACKUP -----'
+    ]
+    assert os.path.isfile(os.path.join(backup_dir, 'xxx', 'iTHX-W3-5_01234.sqlite3'))
+    assert os.path.isfile(os.path.join(backup_dir, 'xxx', 'iTHX-W_56789.sqlite3'))
+    assert os.path.isfile(os.path.join(backup_dir, 'xxx', 'log.txt'))
+    shutil.rmtree(backup_dir)
+
+
+@pytest.mark.skipif(sys.version_info[:2] > (3, 6), reason='not Python 3.6')
+def test_backup_py36():
+    process = run(['omega-logger', 'tests/resources/config_backup.xml', '--backup'], stderr=PIPE, stdout=PIPE)
+    assert process.returncode == 1
+    assert not process.stdout
+    assert process.stderr.rstrip() == b'Can only perform a database backup with Python 3.7 or later'